chore(ui): rename more dumb render components to Base- as per vue convention
