+++ ui/src/components/BaseConfirmDialog.vue
