+++ ui/src/components/BaseEmbedMew.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="q-mx-sm q-my-sm">
-    <MewContent :mew="embedMew.mew as Mew" class="cursor-pointer q-mb-sm" />
+    <BaseMewContent :mew="embedMew.mew as Mew" class="cursor-pointer q-mb-sm" />
     <BaseAgentProfile
       :agent-pub-key="embedMew.action.author"
       :profile="embedMew.author_profile"
@@ -11,7 +11,7 @@
 <script setup lang="ts">
 import { EmbedMew } from "@/types/types";
 import { Mew } from "@/types/types";
-import MewContent from "@/components/MewContent.vue";
+import BaseMewContent from "@/components/BaseMewContent.vue";
 import BaseAgentProfile from "@/components/BaseAgentProfile.vue";
 
 defineProps<{
