+++ ui/src/components/BaseMewListItemSkeleton.vue
